fix transversals

+++ tasks/transversal/include/transversal.hpp
@@ -1,6 +1,6 @@
 #include <cstdint>
 #include <vector>
 
-using TSquare = std::vector<std::vector<uint8_t>>;
+using TSquare = std::vector<std::vector<int>>;
 
 uint64_t transversalNumber(const TSquare& s);
